+++ CHANGELOG.md
@@ -23,8 +23,10 @@
 ### Fixed
 - Fixed checking wrong number of input parameters on methods with array parameters.
 - Fixed use after free of array input arguments ([winsdk#20]).
+- Fixed missing `__await__` on runtime types that inherit `IAsyncOperation` ([winsdk#21]).
 
 [winsdk#20]: https://github.com/pywinrt/python-winsdk/issues/20
+[winsdk#21]: https://github.com/pywinrt/python-winsdk/issues/21
 
 ## [v1.0.0-beta.7] - 2022-09-09